solution #96
Sudoku solver...
This would have been much easier to implement in Prolog -- which I did once before for school. But this was "fun"

Lots of bloat and potential pitfalls in this code. And I'm not sure if its really that readable either. But I am done messing around with sudoku for now!